requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0